requirements.txt
accelerate==0.7.1
albumentations==1.1.0
ants==0.0.7
antspyx==0.3.7
clip==0.2.0
einops==0.4.1
einops_exts==0.0.3
ema_pytorch==0.0.8
grad_cam==1.3.7
h5py==3.6.0
hydra-core==1.1.0
imageio==2.10.3
kornia==0.6.7
monai==1.0.1
natsort==8.2.0
nibabel==3.2.1
numba==0.54.0
numpy==1.20.3
omegaconf==2.1.2
opencv_python==4.5.4.60
pandas==1.3.4
Pillow==9.4.0
python-dotenv==0.21.1
pytorch_lightning==1.5.1
PyYAML==6.0
requests==2.26.0
rich==13.3.1
rotary_embedding_torch==0.1.5
scikit_image==0.18.3
scikit_learn==1.0.2
scipy==1.7.2
seaborn==0.11.2
SimpleITK==2.2.1
timm==0.6.7
torch==1.10.0
torchio==0.18.73
torchmetrics==0.8.1
torchvision==0.11.1
tqdm==4.62.3
transformers==4.26.1
ttach==0.0.3
video_diffusion_pytorch==0.5.3
wandb==0.12.19
matplotlib==3.5.2
hydra-colorlog
